[WIP-vue.js] ui: add topology filters
The user can define a set of filters based on nodes metadatas.

Given a list of filters [(metadata, value), ...] each node that does
not match one of the filters is removed from the d3 graph, including
its links with other nodes.

If the node doesn't have the metadata specified in the filter, the node
is not removed.

Change-Id: I3453c787b973af5d96c788729b1eb1d88d9d180b